+++ packages/apps/doctor/package.json
@@ -9,7 +9,7 @@
     "pack": "node ./scripts/pack.js",
     "prepack-p": "yarn run clean",
     "pack-p": "node ./scripts/pack.js -p",
-    "postpack-p": "ares-package -n ./dist && ares-install --device patient ./com.housepital.app.doctor_1.0.0_all.ipk",
+    "postpack-p": "ares-package -n ./dist ../../js-services/doctorService && ares-install --device doctor ./com.housepital.doctor.app_1.0.0_all.ipk",
     "watch": "node ./scripts/pack.js --watch",
     "clean": "node ./scripts/clean.js",
     "lint": "node ./scripts/lint.js .",
