+++ packages/js-services/doctorService/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "com.housepital.doctor.app.service",
+  "version": "1.0.0",
+  "description": "Helloworld service",
+  "main": "doctor_webos_service.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "BSD"
+}
